+++ pkg/cmd/create/cluster/exec.go
@@ -203,8 +203,13 @@ func (o *Options) runWithClient(kubeClient kubernetes.Interface,
 
 	files = []string{
 		"attach/hub/managed_cluster_cr.yaml",
-		"attach/hub/klusterlet_addon_config_cr.yaml",
 	}
+
+	if helpers.IsRHACM(o.CMFlags) {
+		files = append(files,
+			"attach/hub/klusterlet_addon_config_cr.yaml")
+	}
+
 	out, err = applier.ApplyCustomResources(attachreader, o.values, o.CMFlags.DryRun, "", files...)
 	if err != nil {
 		return err