+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Additions
 
 -  [Add support to AWS clusterpool to support a list of zones #236](https://github.com/stolostron/cm-cli/issues/236)
-
+- [klusteraddon error when create cluster on MCE #240](https://github.com/stolostron/cm-cli/issues/240)
 ## Breaking changes
 
 ## Bug fixes